refactor: move `store.ts` module to `store/db.ts`
